Split routes into submodules
