+++ crates/control/Cargo.toml
@@ -19,6 +19,7 @@ anyhow = "1.0.52"
 axum = "0.4.3"
 chrono = "0.4.19"
 config = "0.11.0"
+futures = "0.3.19"
 hyper = { version = "0.14.16", features = ["full"] }
 once_cell = "1.9.0"
 reqwest = { version = "0.11.9", features = ["json"] }
@@ -33,6 +34,7 @@ tracing = "0.1.29"
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
 
 [dev-dependencies]
+axum-debug = "0.3.2"
 ctor = "0.1.21"
 insta = "1.10.0"
