+++ crates/control/src/controllers/mod.rs
@@ -1,0 +1,2 @@
+pub mod connectors;
+pub mod health_check;
